wb_sdram_ctrl: Fix write and burst cache refill clash
If a write would occur during a burst cache refill and
the write address have not yet been refilled, the
burst cache value would be the one prior to the write.
To fix this, the write is held off until the refill
to that particular address have finished and then do the
write to both the burst cache and the actual memory.