de2_115, de0_nano: phase shift sdram clock
This was suggested by vitorqa on the opencores forums [1]
as a solution for a problem another user had.
As he claims in his post, the Using_the_SDRAM [2]
document from altera suggests shifting the sdram clock
a necessity as well.

I've tested this on the de0 nano and it works fine.

[1] http://opencores.org/forum,OpenRISC,0,5024
[2] ftp://ftp.altera.com/up/pub/Altera_Material/11.0/Tutorials/Verilog/DE0-Nano/Using_the_SDRAM.pdf